+++ src/time/time_slice.rs
@@ -1,12 +1,12 @@
 // 一个时间切片，代表一个时间段
 pub struct TimeSlice {
-    pub id: u128,
-    pub start_time: u128,
-    pub end_time: u128,
+    pub id: u64,
+    pub start_time: u64,
+    pub end_time: u64,
 }
 
 impl TimeSlice {
-    pub fn new(id: u128, start_time: u128, end_time: u128) -> Self {
+    pub fn new(id: u64, start_time: u64, end_time: u64) -> Self {
         Self {
             id,
             start_time,
@@ -15,7 +15,7 @@ impl TimeSlice {
     }
 
     // 获取切片的长度，单位为毫秒
-    pub fn get_len(&self) -> u128 {
+    pub fn get_len(&self) -> u64 {
         self.end_time - self.start_time
     }
 }
